Move FastScanner class to the java/ru/main/polis/ads folder
